Rationalize where all the supporting files live
